@@ -4,7 +4,7 @@ version: 0.0.1
 homepage: https://Mariganesh55@bitbucket.org/Mariganesh55/m2pcarddetails.git
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
@@ -20,6 +20,7 @@ dependencies:
   cryptography: ^2.0.1
   pointycastle: ^3.0.1
   basic_utils: ^3.1.0
+  convert: ^3.0.0
   flutter:
     sdk: flutter 
   
@@ -30,6 +31,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner: ^2.0.4
+  json_serializable: ^4.1.3 
+   
 
 
 # For information on the generic Dart part of this file, see the
