@@ -1,3 +1,3 @@
 class HttpUrl {
-  static const String baseurl = 'https://sit-secure.yappay.in/Yappay/bitUrl/v2';
+  static const String baseurl = 'https://sit-secure.yappay.in/Yappay';
 }
